1.请假详情；2.添加AdminLeave后台控制器
