+++ php/Application/Home/Controller/LeaveNoteController.class.php
@@ -2,7 +2,7 @@
 namespace Home\Controller;
 use Think\Controller;
 class LeaveNoteController extends Controller {
-    
+    // 请假条，所有人都可以看
     public function index(){
         $leaveId = I('id');
         $leaveNote = D('Leave') -> getLeaveById($leaveId);
