+++ web/Admin/common/nav.html
@@ -5,7 +5,7 @@
         <li class="{$userNav}"><a href="{$Think.CONFIG.DOMAIN_URL}/?c=Admin&a=user" role="tab">用户管理</a></li>
         <li class="{$classesNav}"><a href="{$Think.CONFIG.DOMAIN_URL}/?c=Admin&a=classes" role="tab">班级管理</a></li>
         <li class="{$nameNav}"><a href="{$Think.CONFIG.DOMAIN_URL}/?c=Admin&a=name" role="tab">点名管理</a></li>
-        <li class="{$leaveNav}"><a href="{$Think.CONFIG.DOMAIN_URL}/?c=Admin&a=leave" role="tab">请假管理</a></li>
+        <li class="{$leaveNav}"><a href="{$Think.CONFIG.DOMAIN_URL}/?c=AdminLeave&a=leave" role="tab">请假管理</a></li>
         <!-- <li class="{$courseNav}"><a href="{$Think.CONFIG.DOMAIN_URL}/?c=Admin&a=course" role="tab">课程管理</a></li> -->
         <li class="{$termNav}"><a href="{$Think.CONFIG.DOMAIN_URL}/?c=Admin&a=termset" role="tab">学期管理</a></li>
       </ul>
