+++ web/AdminLeave/leaveDetail.html
@@ -7,27 +7,27 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <!-- 上述3个meta标签*必须*放在最前面，任何其他内容都*必须*跟随其后！ -->
     <title>考勤管理系统</title>
-    <link href="./common/headContent.html?__inline" rel="import" />
+    <link href="../Admin/common/headContent.html?__inline" rel="import" />
 </head>
 
 <body>
     <nav class="navbar navbar-default">
         <div class="container-fluid">
             <!-- Brand and toggle get grouped for better mobile display -->
-            <link href="./common/header.html?__inline" rel="import" />
+            <link href="../Admin/common/header.html?__inline" rel="import" />
         </div>
     </nav>
     <div class="container-fluid">
         <div>
             <!-- Nav tabs -->
-            <link href="./common/nav.html?__inline" rel="import" />
+            <link href="../Admin/common/nav.html?__inline" rel="import" />
             <!-- Tab panes -->
             <div class="tab-content">
                 <!-- 点名管理 -->
                 <div role="tabpanel" class="active panel_side tab-pane" id="user">
                     <div>
                         <ul class="nav nav-pills nav-stacked col-md-1" role="tablist">
-                            <li role="presentation" class="active"><a href="{$Think.CONFIG.DOMAIN_URL}/?c=Admin&a=leave">请假记录</a></li>
+                            <li role="presentation" class="active"><a href="{$Think.CONFIG.DOMAIN_URL}/?c=AdminLeave&a=leave">请假记录</a></li>
                         </ul>
                         <div class="tab-content col-md-11">
                             <div role="tabpanel" class="tab-pane active" id="studentList">
@@ -93,7 +93,7 @@
                                                 <div class="col-sm-8 col-sm-offset-1">{$leaveDetail['student']['dorm']}{$leaveDetail['student']['room']}</div>
                                             </div>
                                             <div class="row mt">
-                                                <div class="col-sm-2 text-right">照片</div>
+                                                <div class="col-sm-2 text-right">图片</div>
                                                 <div class="col-sm-8 col-sm-offset-1">
                                                     <if condition="$leaveDetail['image']">
                                                         <img src="{$Think.CONFIG.DOMAIN_URL}/{$leaveDetail['image']}" alt="">
@@ -102,6 +102,34 @@
                                                     </if>
                                                 </div>
                                             </div>
+                                            <div class="row mt">
+                                                <div class="col-sm-2 text-right">学生照片</div>
+                                                <div class="col-sm-8 col-sm-offset-1">
+                                                    <if condition="$leaveDetail['student']['image']">
+                                                        <img src="{$Think.CONFIG.DOMAIN_URL}/{$leaveDetail['student']['image']}" alt="">
+                                                    <else />
+                                                        照片未上传
+                                                    </if>
+                                                </div>
+                                            </div>
+                                        </div>
+                                        <div class="panel-footer">
+                                            <php>
+                                                if ($leaveDetail['status'] == 0) {
+                                            </php>
+                                            <a class="btn btn-success" href="{$Think.config.DOMAIN_URL}/?c=AdminLeave&a=setLeaveAgree&id={$leaveDetail.id}" class="weui_btn weui_btn_mini weui_btn_primary">同意</a>
+                                            <a class="btn btn-danger" href="{$Think.config.DOMAIN_URL}/?c=AdminLeave&a=setLeaveRefuse&id={$leaveDetail.id}" class="weui_btn weui_btn_mini weui_btn_primary">拒绝</a>
+                                            <php>
+                                                } else if ($leaveDetail['status'] == 1) {
+                                            </php>
+                                            <a class="btn btn-danger" href="{$Think.config.DOMAIN_URL}/?c=AdminLeave&a=setLeaveRefuse&id={$leaveDetail.id}" class="weui_btn weui_btn_mini weui_btn_primary">同意转为拒绝</a>
+                                            <php>
+                                                } else {
+                                            </php>
+                                            <a class="btn btn-success" href="{$Think.config.DOMAIN_URL}/?c=AdminLeave&a=setLeaveAgree&id={$leaveDetail.id}" class="weui_btn weui_btn_mini weui_btn_primary">拒绝转为同意</a>
+                                            <php>
+                                                }
+                                            </php>
                                         </div>
                                     </div>
                                 </div>
@@ -112,7 +140,7 @@
             </div>
         </div>
     </div>
-    <link href="./common/footerContent.html?__inline" rel="import" />
+    <link href="../Admin/common/footerContent.html?__inline" rel="import" />
     <script>
     $('.js-search-name-input').on('keyup', function (e) {
         if(e.keyCode==13){
@@ -126,7 +154,7 @@
         //     return;
         // }
 
-        window.location.href = "{$Think.CONFIG.DOMAIN_URL}/?c=Admin&a=leave&keyword=" + $('.js-search-name-input').val();
+        window.location.href = "{$Think.CONFIG.DOMAIN_URL}/?c=AdminLeave&a=leave&keyword=" + $('.js-search-name-input').val();
     });
     </script>
 </body>
