+++ web/AdminLeave/leaveTicket.html
@@ -7,7 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <!-- 上述3个meta标签*必须*放在最前面，任何其他内容都*必须*跟随其后！ -->
     <title>考勤管理系统</title>
-    <link href="./common/headContent.html?__inline" rel="import" />
+    <link href="../Admin/common/headContent.html?__inline" rel="import" />
     <style>
     body {
         font-size: 14px;
@@ -63,7 +63,7 @@
             <li>开始时间</li>
         </ul>
     </div> -->
-    <link href="./common/footerContent.html?__inline" rel="import" />
+    <link href="../Admin/common/footerContent.html?__inline" rel="import" />
 </body>
 
 </html>
